+++ lib/phlex/deferred_render_with_main_content.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Phlex
+  module DeferredRenderWithMainContent
+    def template(&)
+      output = capture(&)
+      super { unsafe_raw(output) }
+    end
+  end
+end